+++ src/main/java/com/summit/service/impl/CameraDeviceServiceImpl.java
@@ -7,6 +7,7 @@ import com.summit.dao.entity.Page;
 import com.summit.dao.repository.CameraDeviceDao;
 import com.summit.dao.repository.LockInfoDao;
 import com.summit.service.CameraDeviceService;
+import com.summit.util.PageConverter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -55,23 +56,39 @@ public class CameraDeviceServiceImpl implements CameraDeviceService {
 
     @Override
     public int delDeviceByDevId(String devId) {
+        if(devId == null){
+            log.error("设备id地址为空");
+            return -1;
+        }
         UpdateWrapper<CameraDevice> wrapper = new UpdateWrapper<>();
         return cameraDeviceDao.delete(wrapper.eq("dev_id", devId));
     }
 
     @Override
     public int delDeviceByIpAddress(String deviceIp) {
+        if(deviceIp == null){
+            log.error("设备ip地址为空");
+            return -1;
+        }
         UpdateWrapper<CameraDevice> wrapper = new UpdateWrapper<>();
         return cameraDeviceDao.delete(wrapper.eq("device_ip", deviceIp));
     }
 
     @Override
     public CameraDevice selectDeviceById(String devId) {
+        if(devId == null){
+            log.error("设备id为空");
+            return null;
+        }
         return cameraDeviceDao.selectDeviceById(devId);
     }
 
     @Override
     public CameraDevice selectDeviceByIpAddress(String deviceIp) {
+        if(deviceIp == null){
+            log.error("设备ip地址为空");
+            return null;
+        }
         CameraDevice cameraDevice = new CameraDevice();
         cameraDevice.setDeviceIp(deviceIp);
         List<CameraDevice> devices = cameraDeviceDao.selectCondition(cameraDevice, null);
@@ -80,6 +97,11 @@ public class CameraDeviceServiceImpl implements CameraDeviceService {
 
     @Override
     public List<CameraDevice> selectDeviceByLockCode(String lockCode, Page page) {
+        if(lockCode == null){
+            log.error("锁编号为空");
+            return null;
+        }
+        PageConverter.convertPage(page);
         CameraDevice cameraDevice = new CameraDevice();
         cameraDevice.setLockCode(lockCode);
         return cameraDeviceDao.selectCondition(cameraDevice, page);
@@ -87,6 +109,7 @@ public class CameraDeviceServiceImpl implements CameraDeviceService {
 
     @Override
     public List<CameraDevice> selectCondition(CameraDevice cameraDevice, Page page) {
+        PageConverter.convertPage(page);
         if(cameraDevice == null){
             log.error("设备信息为空");
             return null;
@@ -96,6 +119,7 @@ public class CameraDeviceServiceImpl implements CameraDeviceService {
 
     @Override
     public List<CameraDevice> selectAll(Page page) {
+        PageConverter.convertPage(page);
         return cameraDeviceDao.selectCondition(new CameraDevice(), page);
     }
